구매 후기 페이지 CSS 수정
