+++ front_end/package.json
@@ -1,6 +1,7 @@
 {
   "name": "front_end",
   "version": "0.1.0",
+  "proxy": "http://localhost:3001",
   "private": true,
   "dependencies": {
     "@react-oauth/google": "^0.12.1",
